Website login system
website - followed a tutorial , it has a login system with hashing for user security aswell as sqlalchamey as the data base for now, can change easily later. same goes for the HTML front end. 

APItesting for the spotify/spotipy started

STILL NEED TO ADD THE VIRTUAL ENVIRONMENT (MAYBE ) it was to ensure the packages were all there however since there is the requirements.txt might be okay without it.